fix/app

+++ app/src/main/java/com/fpoly/smartlunch/data/model/OrderResponse.kt
@@ -7,7 +7,7 @@ data class OrderResponse(
     val userId: User,
     val couponId: String,
     val products: List<ProductOrder>,
-    val deliveryFee: Int,
+    val deliveryFee: Double,
     val discount: Double,
     val total: Double,
     val status: Status,
